fix(curriculum): carry List year filter as int64 to avoid int32 overflow (CodeQL #35)
The List filter's applicable_from_year is untrusted query input. Narrowing it
to int32 wrapped out-of-range values (e.g. 2^31 → negative), which could
spuriously match a stored cohort year. Carry it as sql.NullInt64 with a
$4::bigint cast, mirroring the discipline_id / author_id filters; out-of-range
years now match no row. Removes the conversion CodeQL go/incorrect-integer-conversion
flagged (the prior #nosec only silenced gosec).

+++ internal/modules/work_program/infrastructure/persistence/work_program_repository_pg_test.go
@@ -441,13 +441,13 @@ func TestWorkProgramRepositoryPG_List_FilterByAuthor_PassesAuthorIDArg(t *testin
 	authorID := int64(7)
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT COUNT(*) FROM work_programs")).
 		WithArgs(
-			"", sql.NullInt64{}, "", sql.NullInt32{},
+			"", sql.NullInt64{}, "", sql.NullInt64{},
 			sql.NullInt64{Int64: 7, Valid: true},
 		).
 		WillReturnRows(sqlmock.NewRows([]string{"count"}).AddRow(1))
 	mock.ExpectQuery(regexp.QuoteMeta("FROM work_programs")).
 		WithArgs(
-			"", sql.NullInt64{}, "", sql.NullInt32{},
+			"", sql.NullInt64{}, "", sql.NullInt64{},
 			sql.NullInt64{Int64: 7, Valid: true},
 			20, 0,
 		).
@@ -513,12 +513,12 @@ func TestWorkProgramRepositoryPG_List_FilterByStatusAndSpecialty(t *testing.T) {
 	status := domain.StatusApproved
 	mock.ExpectQuery(regexp.QuoteMeta("SELECT COUNT(*) FROM work_programs")).
 		WithArgs(
-			"approved", sql.NullInt64{}, "09.03.01", sql.NullInt32{}, sql.NullInt64{},
+			"approved", sql.NullInt64{}, "09.03.01", sql.NullInt64{}, sql.NullInt64{},
 		).
 		WillReturnRows(sqlmock.NewRows([]string{"count"}).AddRow(3))
 	mock.ExpectQuery(regexp.QuoteMeta("FROM work_programs")).
 		WithArgs(
-			"approved", sql.NullInt64{}, "09.03.01", sql.NullInt32{}, sql.NullInt64{},
+			"approved", sql.NullInt64{}, "09.03.01", sql.NullInt64{}, sql.NullInt64{},
 			10, 20,
 		).
 		WillReturnRows(sqlmock.NewRows(wpListColumnsTest()).
